+++ src/widgets/dialog_helper.c
@@ -137,13 +137,13 @@ static widget_t* dialog_create_simple_ex(widget_t* parent, xy_t x, xy_t y, wh_t 
 
   widget_set_name(title, DIALOG_CHILD_TITLE);
   widget_set_tr_text(title, stitle);
-  widget_set_self_layout_params(title, "0", "0", "100%", "30");
+  widget_set_self_layout(title, "default(x=0,y=0,w=100%,h=30)");
 
   client = dialog_client_create(widget, 0, 0, 0, 0);
   goto_error_if_fail(client != NULL);
   widget_set_name(client, DIALOG_CHILD_CLIENT);
 
-  widget_set_self_layout_params(client, "0", "bottom", "100%", "-30");
+  widget_set_self_layout(client, "default(x=0,y=bottom,w=100%,h=-30)");
 
   return widget;
 error: